:truck: api: Move pickChallengeService to evaluation
